Created a second model and compared model predictions; wrote an analysis of ways model could be improved on.
note: I would have committed earlier if I had been able to knit, but my analysis document was getting too long to knit properly, causing RStudio to crash. I fixed it once I figured out what was happening, but this is the sum of all the work I've done in the meantime.